chore: move TextInputNode to another file
